correlate_sparse has kernel size check. Updated the formatting.
