+++ skimage/filters/thresholding.py
@@ -11,7 +11,7 @@ from ..util import crop, dtype_limits
 from ..filters._multiotsu import (_get_multiotsu_thresh_indices_lut,
                                   _get_multiotsu_thresh_indices)
 
-from ._sparse import correlate_sparse
+from ._sparse import correlate_sparse, _validate_window_size
 
 
 __all__ = ['try_all_threshold',
@@ -832,26 +832,6 @@ def threshold_triangle(image, nbins=256):
         arg_level = nbins - arg_level - 1
 
     return bin_centers[arg_level]
-
-
-def _validate_window_size(axis_sizes):
-    """Ensure all sizes in ``axis_sizes`` are odd.
-
-    Parameters
-    ----------
-    axis_sizes : iterable of int
-
-    Raises
-    ------
-    ValueError
-        If any given axis size is even.
-    """
-    for axis_size in axis_sizes:
-        if axis_size % 2 == 0:
-            msg = ('Window size for `threshold_sauvola` or '
-                   '`threshold_niblack` must not be even on any dimension. '
-                   'Got {}'.format(axis_sizes))
-            raise ValueError(msg)
 
 
 def _mean_std(image, w):
